Refactor structure and export components
