+++ src/components/task/common/color-bar.js
