+++ src/components/task/task-edit/repeat-block.js
@@ -1,4 +1,4 @@
-import {DAYS, Decision, Checked} from "../../../../consts";
+import {DAYS, Decision, Checked} from "../../../consts";
 
 
 /**
