+++ src/components/tasks-list.js
@@ -1,4 +1,4 @@
-import AbstractComponent from "../abstract/abstract-component";
+import AbstractComponent from "./abstract/abstract-component";
 
 /**
  * Создание разметки блока списка задач
