+++ src/main.js
@@ -2,9 +2,9 @@ import {CountTask, Position, Filter} from "./consts";
 import {generateTasks} from "./mock/tasks/tasks";
 import {generateFilters} from "./mock/filters/filters";
 import {render} from "./utils/change-component";
-import MenuComponent from "./components/menu/menu.js";
-import FiltersComponent from "./components/filters/filters.js";
-import BoardComponent from "./components/board/board.js";
+import {Menu} from "./components/menu.js";
+import {Filters} from "./components/filters.js";
+import {Board} from "./components/board.js";
 import {BoardController} from "./controllers/board.js";
 import {getFilteredTasks, setCheckFilter, addListenersToFilters} from "./controllers/filters";
 
@@ -21,11 +21,11 @@ const Nodes = {
 const init = () => {
   const tasks = generateTasks(CountTask.ALL);
   const filters = generateFilters(tasks);
-  const filtersComponent = new FiltersComponent(filters);
-  const boardComponent = new BoardComponent(getFilteredTasks(tasks));
+  const filtersComponent = new Filters(filters);
+  const boardComponent = new Board(getFilteredTasks(tasks));
   const boardController = new BoardController(boardComponent);
 
-  render[Position.BEFORE_END](Nodes.HEADER, new MenuComponent());
+  render[Position.BEFORE_END](Nodes.HEADER, new Menu());
   render[Position.BEFORE_END](Nodes.MAIN, filtersComponent);
   render[Position.BEFORE_END](Nodes.MAIN, boardComponent);
